gui/requirements.txt
dateutils==0.6.7
PyQt5==5.14.1
PyQt5-sip==12.7.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
